Added RDF format option on GraphDBSUT scripts for repository creation - in most cases avoids the unnecessary conversion of N-TRIPLES to TRIG - speeds up repo creation when dealing with large triple files

+++ GraphDBSUT/scripts/CreateRepos/createGraphDBRepo.sh
@@ -1,10 +1,11 @@
 # SYNTAX :
-#    <script> repoConf ntripleDir graphDBBaseDir
+#    <script> repoConf RDFDir RDFFormat graphDBBaseDir
 SCRIPT_NAME=`basename "$0"`
 SYNTAX="
 SYNTAX: $SCRIPT_NAME repoConf ntripleDir graphDBBaseDir
 \trepoConf\t:\tconfiguration file for repo,
-\tntripleDir\t:\tdirectory with the N-Triple data files to load,
+\tRDFDir\t:\tdirectory with the RDF data files to load,
+\tRDFFormat\t:\tRDF format {N-TRIPLES | TRIG }
 \tgraphDBBaseDir\t:\tbase directory of the GraphDB installation"
 
 # Assumptions
@@ -23,7 +24,7 @@ MAP_CONTEXTS_FILE=$BASE/$MAP_CONTEXTS_FILE
 
 # STEP 1: Validate the script's syntax
 #      1.1: check number of arguments and assign them to variables
-if (( $# != 3 )); then
+if (( $# != 4 )); then
     echo -e "Illegal number of parameters $SYNTAX"
 	exit 1
 fi
@@ -35,7 +36,9 @@ RepoName=`grep -e "repositoryID" $RepoConfig | awk -F"\"" ' { printf $2 }'`
 # echo $RepoName
 NtripleDir=$2
 # echo $NtripleDir
-GraphDBBaseDir=$3
+RDFFormat=${3^^}
+# echo $RDFFormat
+GraphDBBaseDir=$4
 # echo $GraphDBBaseDir
 
 #      1.3: check whether the directories (ntripleDir graphDBBaseDir) exist
@@ -47,12 +50,18 @@ for dir in "${dirs[@]}"; do
 	fi		
 done
 
+#      1.4: check whether the RDF format is N-Triples or TRIG
+if [ "${RDFFormat}" != "TRIG" ] && [ "${RDFFormat}" != "N-TRIPLES" ]; then
+    echo "Supported RDF formats are : {N-TRIPLES | TRIG}!"
+    exit 3
+fi
+
 # STEP 2: Assess whether the script's preconditions are met
 #      2.1: check whether the $GraphDBBaseDir/bin/preload exists
 PreLoad_Exe="${GraphDBBaseDir}/bin/preload"
 if [ ! -e "$PreLoad_Exe" ]; then
 	echo "File $PreLoad_Exe does not exist!"
-	exit 3
+	exit 4
 fi
 
 #      2.2: check whether the repos $RepoName already exists
@@ -68,7 +77,7 @@ repoDir="${GraphDB_Data_Dir}/repositories/${RepoName}"
 # echo $repoDir 
 if [ -d "$repoDir" ]; then
 	echo "Repo $RepoName already exists in --> $repoDir"
-	exit 4
+	exit 5
 fi		
 
 # STEP 3: Terminate the GraphDB server process, if it's running
@@ -77,36 +86,44 @@ kill -9 `pgrep -f graphdb` > /dev/null 2>&1
 # STEP 4: For the directory $NtripleDir create TRIG from N-Triple files
 #      4.1: change working directory to $NtripleDir
 cd $NtripleDir
-#      4.2: check if $MAP_CONTEXTS_FILE file exists
-if [ -e $MAP_CONTEXTS_FILE ]; then
-	MapToContextFile_Exists=1
+#      4.2: if $RDFFormat is TRIG, convert N-Triples to TRIG
+if [ "${RDFFormat}" = "TRIG" ]; then
+    #      4.2.1: check if $MAP_CONTEXTS_FILE file exists
+    if [ -e $MAP_CONTEXTS_FILE ]; then
+            MapToContextFile_Exists=1
+    else
+            MapToContextFile_Exists=0
+    fi
+    #      4.2.2: For each N-Triple file in the $datafir do ...
+    for i in *.nt; do 
+    #      4.2.3: convert N-Triple file to TRIG with default graph using the rdf2rdf-1.0.1-2.3.1.jar program
+            filename=$(basename "$i"); 
+            extension="${filename##*.}"; 
+            fname="${filename%.*}"; 
+            trigfilename="${fname}.trig"; 
+            java -jar "${BASE}/rdf2rdf-1.0.1-2.3.1.jar" $filename $trigfilename;
+    #      4.2.4: if $MapToContextFile_Exists set the corresponding graph IRI in the TRIG file
+            if [ $MapToContextFile_Exists -eq 1 ]; then
+                    matchedline=`grep -e $i $MAP_CONTEXTS_FILE`
+                    # echo "File $i found in file $MAP_CONTEXTS_FILE in line : $matchedline"
+                    matchedcontext=`echo -e "$matchedline" | awk -F"\t" ' { printf $2 }'`
+                    # echo "Corresponding context is : $matchedcontext"
+                    sed -i 's+{$+'${matchedcontext}' {+' $trigfilename
+                    echo "Modified graph name in TRIG file $trigfilename to $matchedcontext"
+            fi
+    done
+    #      4.2.5: Set the filetype to TRIG
+    rdffiletype='.trig'
 else
-	MapToContextFile_Exists=0
+    #      4.2.6: Set the filetype to N-TRIPLES
+    rdffiletype='.nt'
 fi
-#      4.3: For each N-Triple file in the $datafir do ...
-for i in *.nt; do 
-#      4.3.1: convert N-Triple file to TRIG with default graph using the rdf2rdf-1.0.1-2.3.1.jar program
-	filename=$(basename "$i"); 
-	extension="${filename##*.}"; 
-	fname="${filename%.*}"; 
-	trigfilename="${fname}.trig"; 
-	java -jar "${BASE}/rdf2rdf-1.0.1-2.3.1.jar" $filename $trigfilename;
-#      4.3.2: if $MapToContextFile_Exists set the corresponding graph IRI in the TRIG file
-	if [ $MapToContextFile_Exists -eq 1 ]; then
-		matchedline=`grep -e $i $MAP_CONTEXTS_FILE`
-		# echo "File $i found in file $MAP_CONTEXTS_FILE in line : $matchedline"
-		matchedcontext=`echo -e "$matchedline" | awk -F"\t" ' { printf $2 }'`
-		# echo "Corresponding context is : $matchedcontext"
-		sed -i 's+{$+'${matchedcontext}' {+' $trigfilename
-		echo "Modified graph name in TRIG file $trigfilename to $matchedcontext"
-	fi
-done
-
-# STEP 5: Create $RepoName repository with config file $RepoConfig and load TRIG files while measuring the elapsed time
-time $PreLoad_Exe -c $RepoConfig *.trig
+# STEP 5: Create $RepoName repository with config file $RepoConfig and load RDF files while measuring the elapsed time
+time $PreLoad_Exe -c $RepoConfig *$rdffiletype
 
 # STEP 6: Remove TRIG file to avoid allocating space
-rm *.trig
+if [ "${RDFFormat}" = "TRIG" ]; then
+    rm *.trig
+fi
 
 exit 0
-
